feat(hera): Rename bin to hera and allow standalone bin
