+++ Cargo.toml
@@ -10,10 +10,10 @@ categories = ["cryptography", "cryptography::cryptocurrencies"]
 
 [workspace]
 members = [
-  "bin/kona-exex",
+  "bin/hera",
   "crates/kona-providers",
 ]
-default-members = ["bin/kona-exex"]
+default-members = ["bin/hera"]
 
 # Explicitly set the resolver to version 2, which is the default for packages with edition >= 2021
 # https://doc.rust-lang.org/edition-guide/rust-2021/default-cargo-resolver.html
