+++ bin/hera/src/hera.rs
@@ -1,9 +1,54 @@
-//! CLI Extension module for Kona
+//! Module for the Hera CLI and its subcommands.
 
-use std::path::PathBuf;
-
-use clap::Args;
 use url::Url;
+use std::sync::Arc;
+use std::path::PathBuf;
+use eyre::{bail, Result};
+use reth::cli::Cli;
+use reth_exex::ExExContext;
+use reth_node_api::FullNodeComponents;
+use superchain_registry::RollupConfig;
+use clap::{Parser, Args, Subcommand};
+use reth_node_ethereum::EthereumNode;
+use superchain_registry::ROLLUP_CONFIGS;
+
+/// The top-level Hera CLI Command
+#[derive(Debug, Parser)]
+#[command(author, about = "Hera", long_about = None)]
+pub struct HeraCli {
+    /// Hera's subcommands
+    #[command(subcommand)]
+    pub subcmd: HeraSubCmd,
+}
+
+impl HeraCli {
+    /// Runs the Hera CLI
+    pub fn run(self) -> Result<()> {
+        match self.subcmd {
+            HeraSubCmd::ExEx(cli) => cli.run(|builder, args| async move {
+                let Some(cfg) = ROLLUP_CONFIGS.get(&args.l2_chain_id).cloned() else {
+                    bail!("Rollup configuration not found for L2 chain ID: {}", args.l2_chain_id);
+                };
+                let node = EthereumNode::default();
+                let kona = move |ctx| async { Ok(HeraExEx::new(ctx, args, Arc::new(cfg)).await.start()) };
+                let handle = builder.node(node).install_exex(crate::EXEX_ID, kona).launch().await?;
+                handle.wait_for_node_exit().await
+            }),
+            HeraSubCmd::Bin => unimplemented!(),
+        }
+    }
+}
+
+/// The Hera subcommands
+#[derive(Debug, Subcommand)]
+pub enum HeraSubCmd {
+    /// The Execution Extension
+    #[clap(name = "exex")]
+    ExEx(Cli<HeraArgsExt>),
+    /// A standalone rollup node binary.
+    #[clap(name = "bin")]
+    Bin,
+}
 
 /// The default L2 chain ID to use. This corresponds to OP Mainnet.
 pub const DEFAULT_L2_CHAIN_ID: u64 = 10;
@@ -14,8 +59,9 @@ pub const DEFAULT_L2_RPC_URL: &str = "https://optimism.llamarpc.com/";
 /// The default L1 Beacon Client RPC URL to use.
 pub const DEFAULT_L1_BEACON_CLIENT_URL: &str = "http://localhost:5052/";
 
+/// The Hera Execution Extension CLI Arguments.
 #[derive(Debug, Clone, Args)]
-pub(crate) struct KonaArgsExt {
+pub(crate) struct HeraArgsExt {
     /// Chain ID of the L2 network
     #[clap(long = "kona.l2-chain-id", default_value_t = DEFAULT_L2_CHAIN_ID)]
     pub l2_chain_id: u64,
@@ -75,5 +121,28 @@ impl std::str::FromStr for ValidationMode {
             "engine-api" => Ok(ValidationMode::EngineApi),
             _ => Err(format!("Invalid validation mode: {}", s)),
         }
+    }
+}
+
+/// The Hera Execution Extension.
+#[derive(Debug)]
+#[allow(unused)]
+pub(crate) struct HeraExEx<Node: FullNodeComponents> {
+    /// The rollup configuration
+    cfg: Arc<RollupConfig>,
+    /// The context of the Execution Extension
+    ctx: ExExContext<Node>,
+}
+
+#[allow(unused)]
+impl<Node: FullNodeComponents> HeraExEx<Node> {
+    /// Creates a new instance of the Hera Execution Extension.
+    pub async fn new(ctx: ExExContext<Node>, args: HeraArgsExt, cfg: Arc<RollupConfig>) -> Self {
+        Self { ctx, cfg }
+    }
+
+    /// Starts the Execution Extension loop.
+    pub async fn start(mut self) -> Result<()> {
+        unimplemented!()
     }
 }
